Fix conflicts move audio to adapters types
