+++ dspy/adapters/__init__.py
@@ -1,9 +1,7 @@
 from dspy.adapters.base import Adapter
 from dspy.adapters.chat_adapter import ChatAdapter
 from dspy.adapters.json_adapter import JSONAdapter
-from dspy.adapters.types import Image, History, encode_image, is_image
-from dspy.adapters.audio_utils import Audio, encode_audio, is_audio
-from dspy.adapters.media_utils import try_expand_media_tags
+from dspy.adapters.types import Image, History, Audio
 
 __all__ = [
     'Adapter',
@@ -11,10 +9,5 @@ __all__ = [
     'JSONAdapter',
     'Image',
     'Audio',
-    'encode_image',
-    'encode_audio',
-    'is_image',
-    'is_audio',
-    'try_expand_media_tags',
     "History",
 ]
