+++ dspy/adapters/json_adapter.py
@@ -14,7 +14,7 @@ from pydantic.fields import FieldInfo
 
 from dspy.adapters.base import Adapter
 from dspy.adapters.types.image import Image
-from dspy.adapters.audio_utils import Audio
+from dspy.adapters.types.audio import Audio
 from dspy.adapters.types.history import History
 from dspy.adapters.utils import format_field_value, get_annotation_name, parse_value, serialize_for_json
 from dspy.signatures.signature import SignatureMeta
