+++ dspy/adapters/types/__init__.py
@@ -1,4 +1,5 @@
 from dspy.adapters.types.history import History
 from dspy.adapters.types.image import Image
+from dspy.adapters.types.audio import Audio
 
-__all__ = ["History", "Image"]
+__all__ = ["History", "Image", "Audio"]
