+++ dspy/signatures/signature.py
@@ -28,7 +28,7 @@ from pydantic import BaseModel, Field, create_model
 from pydantic.fields import FieldInfo
 
 from dspy.adapters.types.image import Image  # noqa: F401
-from dspy.adapters.audio_utils import Audio  # noqa: F401
+from dspy.adapters.types.audio import Audio  # noqa: F401
 from dspy.signatures.field import InputField, OutputField
